backend/pyproject.toml
[tool.poetry]
name = "boosteropenningbackend"
version = "0.1.0"
description = "The backend of the Booster Openning project"
authors = ["Armand \"Skengrek\" Massé <armand.masse69@gmail.com>"]
license = "gpl-3.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.4"
asgiref = "3.5.2"
certifi = "2022.12.7"
charset-normalizer = "2.1.1"
coreapi = "2.3.3"
coreschema = "0.0.4"
cryptography = "39.0.0"
dacite = "1.6.0"
defusedxml = "0.7.1"
django = "4.1.3"
django-cors-headers = "3.13.0"
django-templated-mail = "1.1.1"
djangorestframework = "3.14.0"
djangorestframework-simplejwt = "4.8.0"
idna = "3.4"
itypes = "1.2.0"
jinja2 = "3.1.2"
markupsafe = "2.1.1"
oauthlib = "3.2.2"
pokemontcgsdk = "3.4.0"
pycparser = "2.21"
pyjwt = "2.6.0"
python3-openid = "3.2.0"
pytz = "2022.6"
requests = "2.28.1"
requests-oauthlib = "1.3.1"
six = "1.16.0"
social-auth-app-django = "4.0.0"
social-auth-core = "4.3.0"
sqlparse = "0.4.3"
uritemplate = "4.1.1"
urllib3 = "1.26.13"
debugpy = "^1.6.7"
python-dotenv = "^1.0.0"
tqdm = "^4.66.1"
cffi = "^1.17.0"
pillow = "^10.4.0"
setuptools = "^72.2.0"
psycopg2 = "^2.9.9"
django-formtools = "^2.5.1"
celery = "^5.4.0"
redis = "^5.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
